+++ components/demo/GetLoggingInStateDemo.tsx
@@ -1,6 +1,7 @@
 import { Text } from '@chakra-ui/react';
 import { FlexCardWrapper } from '../ui/CardWrapper';
 import { useLoggingIn } from '../../hooks/auth/useLoggingIn';
+import { CardItemWrapper } from './CardItemWrapper';
 
 export const GetLoggingInStateDemo = () => {
   const { isLoggingIn, error, isLoggedIn } = useLoggingIn();
@@ -10,24 +11,24 @@ export const GetLoggingInStateDemo = () => {
       <Text fontSize="xl" mb={2} fontWeight="black">
         Logging in current state:
       </Text>
-      <Text>
+      <CardItemWrapper>
         <Text as="span" display="inline-block" fontWeight="bold">
           isLoggingIn:
         </Text>{' '}
         {isLoggingIn ? 'true' : 'false'}
-      </Text>
-      <Text>
+      </CardItemWrapper>
+      <CardItemWrapper>
         <Text as="span" display="inline-block" fontWeight="bold">
           error:
         </Text>{' '}
         {error || '-'}
-      </Text>
-      <Text>
+      </CardItemWrapper>
+      <CardItemWrapper>
         <Text as="span" display="inline-block" fontWeight="bold">
           isLoggedIn:
         </Text>{' '}
         {isLoggedIn ? 'true' : 'false'}
-      </Text>
+      </CardItemWrapper>
     </FlexCardWrapper>
   );
 };