+++ hooks/auth/useLoginInfo.tsx
@@ -1,8 +1,8 @@
-import { useSnapshot } from 'valtio';
+import { useProxy } from '../tools/useProxy';
 import { loginInfoState } from '../../store/auth';
 
 export const useLoginInfo = () => {
-  const loginInfoSnap = useSnapshot(loginInfoState);
+  const loginInfoSnap = useProxy(loginInfoState);
 
   return {
     loginMethod: loginInfoSnap.loginMethod,